:recycle: Refactor to stream creation
t
